Consolidate resolve logic and extract source_file_coord.py
- Unify var resolution into resolve_data_var() on each coord subclass:
  OpenData applies lead-time param overrides, MARS merges mars overrides
- Remove resolve_grib_index_params, _resolve_vars_for_coord, _with_resolved_attrs
- Move source file coord classes, constants, and type alias to source_file_coord.py
